Changes to filesystem delete

+++ src/filesystem/FileSystemManager.java
@@ -173,11 +173,12 @@ public class FileSystemManager {
     public boolean deleteBackupFile(String fileID) {
         File dir = new File(this.getPathToBackups(), fileID);
 
-        for (File f : dir.listFiles()) {
-            f.delete();
+        if (dir.exists()) {
+            for (File f : dir.listFiles()) {
+                f.delete();
+            }
+            dir.delete();
         }
-
-        dir.delete();
 
         return true;
     }
